Codebase Structure Revision & Remote Keyboard Control
1. Re-organized the codebase structure (details in readme)
2. Remote control using keyboard (with image shooting command)

+++ auto_car/firmware/encoder.h
@@ -2,6 +2,6 @@
 #define ENCODER_H
 
 void encoderSetup();
-int getEncoderCount();
+long getEncoderCount();
 
 #endif
